Add to-do list
Add a file to track the to-do items for this project. Mostly a lot of
things that I want to do, but am setting aside until I can get the core
functionality up and running.

Anyone who wants to help is welcome to pitch in and send me patches by
email or pull requests on Github.

Signed-off-by: Grant Likely <grant.likely@secretlab.com>