Wrap VISCA API with an object
Wrap the VISCA camera API with an object in preparation of having
multiple PTZ devices. The interface (serial port) portion is still hard
coded, but each PTZ device on the daisychain are wrapped in a PTZCamera
object (name of object may yet change). This will make it easier to
instance multiple PTZ objects.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>