Add support for multiple cameras
Store all the PTZCamera instances in a vector and allow each camera to
be controlled in turn. Adds two buttons, "Next Camera" and "Previous
Camera". Clicking one of these buttons will move forward or backward in
the list of cameras.

Interface to the camera is still fixed to a single serial port. The
plugin will scan the serial port at start up and create a camera
instance for each camera that it finds.

Todo: create a better user interface for selecting cameras

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>