visca: Move interface setup code into PTZVisca class
The common code doesn't need to know the details of the VISCA interface.
All of that detail can be encapsulated in the PTZVisca class. Add a
static map to PTZVisca to keep track of all the open VISCAInterface_t
UARTs and have the PTZVisca constructor just take the name of the UART
as the argument.

Signed-off-by: Grant Likely <grant.likely@secretlab.com>