Populate ListView with the camera names
Add a QStringListModel to PTZDevice that can be used to populate the
list of cameras in the ListView in the control dock. With this commit
the cameras are only listed. It doesn't yet provide any feedback on
which camera is active, or allow the camera to be selected. Those
features will be added soon.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>