Remove pantilt direction wrappers
Separate methods for each direction of pan/tilt is awkward and adds a
lot of duplicate code. If instead the direction of pan/tilt is encoded
into the argument by using a signed pan/tilt speed value, all the
duplicated functions can be removed and the code will be simpler.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>