Add basic scene tracking support
Really simple implementation of scene following. The target of the PTZ
controls can be set to follow scene switches with the radio buttons at
the bottom of the panel. If the selection is set to "Follow Preview" or
"Follow Program", then the PTZ device with the same name as the preview
or program scene will be made active.

This is a fairly unsophisticated approach that can certainly be
improved, but it is a good starting point.

Signed-off-by: Grant Likely <grant.likely@secretlab.com>