Move management of device list into PTZDevice
Move the device list from the PTZControls class to PTZDevice. It makes
more sense for the list of devices to be managed in PTZDevice so that it
can be accessed from multiple views. This also paves the way for the Qt
model to be directly populated by the list of devices.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>