Integrate libvisca into source tree
There is very little code within the libvisca library, and this plugin
is hitting the limit of what libvisca can provide. Merge the libvisca
code into this plugin to simplify the build and so that the libvisca
code can be adapted to what will work best here without breaking
backwards compatibility.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>