Add command queue to PTZVisca
The VISCA protocol only allows one command to be in flight at a time so
that responses can be matched with the command. Add a command queue in
the form of a QByteArrayList to PTZVisca, and put all commands into the
queue. When a response is received, use it to trigger sending the next
command.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>