Add timeout timer for VISCA commands
If a response is not received for a VISCA command, clear the command
buffer so that the camera does not get stuck in the 'frozen' state.
Without this change a missed response will cause control of the camera
to entirely freeze and no more commands will be sent, even if the camera
is reconnected.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>