Create protocol objects for VISCA messages
Create ViscaCmd class to encode VISCA messages and the encoded result.
Knowing about the outgoing message is needed to know how to interpret
incoming messages. ViscaCmd contains the cmd message as well as both the
outgoing and incoming argument encoding, and can be consisely defined in
a single line.

This patch removes all the open coded message parsing and replaces it
with the ViscaCmd object model. It also removes the remaining libvisca
code. This patch lays the groundwork for generic parsing code of
incoming parameters.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>