Use OBSData wrapper to fix obs_data memory leaks
Configuration data stored in obs_data structures is not correctly
reference counted and therefore not freed when OBS exits. Fix this
memory leak by switching to the OBSData wrapper for managing reference
counts to the structures, and add the appropriate calls to
obs_data_release() on structures returned from the C api.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>