ptz-device: Make OBSData structure transient
set_config and get_config is used to save & restore to the configuration
file, and for the settings dialog to retreive the configuration. In both
cases the data is transient and doesn't need to be kept around. Rework
set/get_config() to generate the OBSData structure as needed, and
discard it when complete.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>