Define the reset of the VISCA command set
Adds the rest of the VISCA command set from the SRG-120DH manual. This
probably isn't an exhaustive list of VISCA commands, but it is enough to
show patterns, such as how set and get methods are somewhat related.
Also rearrange the definitions so that set/get operations are grouped
together.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>