Drop 'const' from ViscaCmd members to fix Linux build
Compilation on Linux was failing because GCC didn't want to create a
default operator= assignment operator with const QList<T> members.
Dropping the const fixes it.

There is probably a better solution for this, but the ViscaCmd class is
due for a rewrite anyway to organize sets of commands into operations on
a property (e.g., the pan position could be 'driven' left or right, set
to a specific values, adjusted with an offset, or queried). So this fix
will do in the meantime.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>