VISCA: Factor out common send_pending() code
Much of send_pending() is common between the serial and IP variants of
the VISCA implementation. Factor them out into a single base class
send_pending() so that the code for putting a message on the wire is
isolated. This is preparation work ahead on handing inquiries
differently from commands. In future work commands will be sent
immediately without queueing, while inquiries still need to be queued so
that responses can be matched to the inquiry.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>