VISCA: send commands immediately without queueing
VISCA commands, unlike enquiries, can be sent immediately without
waiting for acknowledgement because the return value doesn't need to be
matched to the specific outgoing command. Change send function to send
commands immediately, but still queue inquiries. This will make the
controls more responsive because there isn't a time lag to sending the
command to the camera.

This also fixes the problem of not sending the stop command if the
camera has not responded. Now the commands get sent unconditionally,
even if the camera is not attached.

Fixes: #7
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>