Fix unable to change settings when only one device
If there is only one camera device, it is impossible to change the
device settings in the settings dialog. The workaround required adding
a second unused device so that the selection could be changed, causing
the dialog to refresh.

Fix the problem be explicitly clearing the selection when the settings
dialog is opened so that the first click on the device will cause a
change event and refresh the dialog data. Before this patch the device
was already selected, which meant the selection changed event wasn't
getting generated.

Fixes: #9
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>