Use git describe to get plugin version
Call git describe from CMAKE to generate the version string and embed it
into the plugin. At the same time rework the settings dialog to
dynamically set the description label instead of being statically set in
the settings.ui file so that the version number can be embedded, and so
that additional information can be added to it at runtime if needed.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>