Add keyboard modifiers to pan/tilt actions
If <ctrl> or <shift> are held when a pan/tilt button is pressed, then
use it to adust the speed. <ctrl> gives fastest pan/tilt speed. <shift>
gives slow single step movements.

This patch also rescales the pan/tilt speed values. Instead of being
scaled to the VISCA specification, use the range [-1.0, 1.0] and scale
the value in the PTZ backend.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>