Add 'ptz_debug()' macro for debugging
Adds the ptz_debug() macro for debug messages in this plugin. The log
level is determined by the value of 'ptz_debug_level', which is loaded
from the configuration file. This allows the log level to be changed
without recompiling. Normally debug messages should not be output to the
log, but it is useful to be able to turn it on. Currently the log level
defaults to LOG_INFO because the plugin is still under heavy
development, but it will be changed to LOG_DEBUG when things settle
down.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>