Fix gamepad settings so it can be turned off
Gamepad control was hard coded to on, with no ability to disable it in
the settings dialog. The checkbox did nothing. Add an item to the
configuration file that allows the gamepad to be enabled/disabled, and
wire it up to the settings dialog.

Also removes the gamepad selection combo box from settings because it
doesn't do anything yet.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>