settings: Only show config settings relevant for device
All of the settings (serial port, ip address, etc) were getting show to
the user regardless of whether or not they were relevant. Fix this by
putting the settings widgets into groups that can be easily
shown/hidden. In the currentChanged() slot show only the group that is
relevant for the selected device.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>