Handle responses from Cisco PrecisionHD cameras
Not all cameras use slots for multiple commands in flight at once. Some
Cisco PrecisionHD models appear to do this. Handle the case of an empty
reply coming back without an acknowledgement first. If an empy reply is
recieved, assume it is the result of a command, and don't try to process
it as an inquiry result.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>