Fix endless resend of failed command
The VISCA driver was endlessly resending commands if the camera returns
an error. This was because the 'active' flag was cleared on receipt of
an error, but the failed command was not cleared out of the queue. Fix
this by dropping the active command when an error message is received.

This patch should fix the problem, but there are edge cases if a control
command and inquiry are both in flight at the same time. The current
VISCA code always sends commands immediately to reduce latency on the
assumption that the camera can handle it, but that may not be the best
approach. This code will need to be revisited at some point to make it
more robust.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>