Don't clear inquiry queue on timeout
If a command times out, the camera may just not have been ready. Don't
clear the entire inquiry queue when that happens. It would possibly be
better to retry a timed out command one or two more times, but for now
just drop it and go on to the next.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>