serial: create get/set accessors for baudRate
The baud rate setting code was implemented outside the UART wrapper
class which required the VISCA and Pelco protocol classes to have direct
access to the baud_rate variable and to directly handle to close() and
open() calls required to change the baud rate. This commit moves the
baud rate handling code into a set of accessor functions so that the
details can be hidden from the protocol classes.

This commit also removes the baud_rate member variable because it isn't
needed. The UART class stores the baud rate setting, and can be
retrieved at any time using uart.baudRate().

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>