Create PTZCmd/PTZInq from ViscaCmd & ViscaInq
The structure of ViscaCmd and ViscaInq is generally useful for any
command protocol, not just VISCA. Move the implementations out of
ptz-visca.hpp and into the common protocol-helpers.hpp file and rename
to PTZCmd & PTZInq so that they can be used by any protocol.

At the same time move all the visca specific datagram formats out of the
header file and into ptz-visca.cpp as that code isn't used anywhere
outside of that file.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>