protocol: decode directly to obs_data structure
Each field in a protocol message can have a different data type. Instead
of decoding everything to an int, pass an obs_data structure into the
decode method for each field so that the correct type can be encoded
into the data.

Also adds support for boolean values when auto generating the properties
display.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>