settings: Don't overwrite preset or name changes
When loading the device settings, remove the name and preset items from
the obs_data structure when populating the OBSPropertiesView so that
when applying configuration changes they don't overwrite other changes
that have been made outside the propertiesView. This is fine because the
properties view isn't used to manage either the name or the list of
presets.

Fixes: #35

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>