ui: enforce aspect ratio of pan/tilt/zoom/focus buttons
Use an event filter to ensure that the aspect ration on buttons remains
square. The event filter checks for resize events, and when one is found
it updates the minimumHeigth to match the width. It also resizes the
button icon to match the available space in the icon.

Also reorganizes the control layout to work with the event filter
scheme. The zoom controls have been moved to the right of the pan/tilt
buttons and added into the grid layout so that width of all the buttons is
even.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>