Move PTZDevice instance list into the PTZListModel
Drop the static member functions from PTZDevice, and move all the data
into PTZListModel so that the list of devices is separated from the
device instances themselves. This does a few nice things; it allows the
Qt list model to be kept with the data, it gives a clearer division
between instances and the management of them, and it provides a way to
call list methods through the Qt event system without figuring out a way
to call static methods through QMetaObject::invokeMethod(). Being able
to use invokeMethod() is particularly helpful for sending PTZ commands
from separate threads, like when calling in from a different OBS plugin.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>