Make PTZ actiion proc handlers thread safe
Rework the proc handlers to use QMetaObject::invokeMethod() so that they
can be called from any thread. If called from the same thread then they
behave as normal (run to completion), but if called from another thread
as they would from a non-UI thread plugin, then the call is a
non-blocking enqueue into the UI thread event loop so that it can be
completed in the right context.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>