Emit a signel when PTZDevice settings change
Adds code to emit a signal whenever the camera settings change, usually
in response to an inquiry packet being received. The control panel UI
can use the signal to update the visible controls. For example, in the
current code if the autofocus status chagnes, then some of the buttons
can be enabled or disabled.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>