Rename "camera" to "device_id" in the action filter
Minor cleanup to the action filter code to rename the `camera` data
element to `device_id`, and rename the camera property to match. This
change is to prepare for breaking the assumption that the device id
equals the array offset in the configuration data.

Warning: Breaking change. Old action source configuration will need to
be reset after this patch.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>