Always use .currCamera() to get ptz device
Retrieving the ptz device pointer was open coded in several places.
Replace all of them with a call to the currCamera() method to simplify
the code and at the same time rely on the camera list view to tell use
what the current camera is instead of caching it in the curr_camera
member variable.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>