Allow device to be associated with any source name
Instead of matching only against the scene name, use
obs_source_enum_active_sources() to check the names of all active
sources in either the program or preview scenes. That way the PTZ device
can be associated with a camera source, and that source may be used in
multiple scenes. With the old method a PTZ device could only be
associated with a single scene which doesn't match how cameras are
commonly used.

Fixes: #3
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>