visca-over-tcp: Automatically reconnect when sending
When sending a command, if the socket is disconnected try to
automatically reconnect at that time. This should solve the problem of
the controls becoming unresponsive because the connection was lost.

This isn't a complete fix. A better solution would be to have a time
periodically trying to reconnect so that the first command after
disconnection doesn't get lost.

Fixes: #57
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>