Only send stop commands when needed
The control dock was unconditionally sending stop commands every time
the selected device chagned which manifested as UI lag of about 1s every
time the scene changed. The lag issue on sending commands is an issue,
and needs to be investigated. However in the mean time the control doc
can be modified to only send stop commands if the PTZ device is actually
in motion.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>