Fix guard against live moves
Occasionally sources are reporting as being active when they aren't.
This causes the guard against live moves to get activated when it
shouldn't. This fix replaces the call to obs_source_active() with a
block that scans the current active scene for the relevant source.

This is a workaround and will probably get reverted if the problem gets
fixed in upstream.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>