gamepad: Make gamepad support configurable
Factor out all gamepad support and put behind a CONFIG_USE_GAMEPAD
preprocessor #define so that it can be compiled out. QGamepad does not
work on Windows, and won't be available initially on MacOS, so make it
possible to disable it entirely from the build.

This also refactors the gamepad support to hold a persistant
QGamepadManager instance and use the connectedGamepadsChanged signal to
look for new gamepad connections.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>