Simplify zoom and focus move apis
Instead of separate functions for zoom_{tele,wide,stop} and
focus_{near,far,stop}, create a single method for each action. Where the
speed parameter is 0.0 for stop, positive value for zoom_tele and
focus_far, and negative value for zoom_wide and focus_near. This
simplifies calling interface which will make it easier to export it as a
proc_handler API.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>