Revert ptz_get_proc_handler() API declaration change
Under review it was suggested that the ptz_get_proc_handler() should be
registered as "void ptz_get_proc_handler(out ptr return)", but 'return'
is a reserved name. The above registration results in the following
error:

```
07:01:10.694: [ptz-controls] PTZ controller OBS Plugin"v0.10.3"
07:01:10.694: Errors/warnings for 'void ptz_get_proc_handler(out ptr return)':
07:01:10.694:
07:01:10.694: declaration (1, 35): 'return' is a reserved name
07:01:10.694:
07:01:10.694: Function declaration invalid: void ptz_get_proc_handler(out ptr return)
```

Fix this by reverting to the original registration:
`ptr ptz_get_proc_handler()`.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>