Prototype ONVIF support
This code is prototype and currently disabled until changes are made. The big
issue is that the code blocks with msleep() or a custom event loop in a few
places which will block the OBS main loop.

This commit squashes together the following commits from the original author:
- Basic implementation of Onvif standard
- Fix reversed pan
- Setting int for absolute and relative functions
- Removed settingsChanged dummy function
- Implemented presets
- Fixes on presets identification

I've also fixed CMakeLists.txt to be able to find the Qt XML libraries, and
cleaned up the minor code convention and structure issues. The code builds now,
but the blocking issues mentioned above need to be addressed before it can be
enabled for general use.

Signed-off-by: Jonatã Bolzan Loss <jonata@jonata.org>
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>