Import obs-plugintemplate helper files
Brings in the unmodified files that aren't used yet, but will be when this
pluging is updated to match the current plugin template. This commit should not
change any of the build behaviour, but it prepares the ground for reworking
CMakeList.txt and main.yml to use the template code.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>