imported: Update OBSPropertiesView widget
Imports the source code for OBSPropertiesWidget from obs-studio proper.
Related code files are based on commit 28.0.0-beta1-56-g43a49dca4 in the
OBS Studio source tree.

As before, this is a suboptimal solution. It would be easier if OBS
studio exported an API for creating the OBSPropertiesWidget and hook up
the callbacks for interacting with it. If I can make that change work
and it gets accepted, then all this code can be pulled out of the tree.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>