Stop dereferencing QMap by '+' operator
QT6 doesn't allow QMap elements to be dereferenced by arithmetic offset because
they are not random access. Instead, get the list of keys and dereference that
by index. This may not actually be correct, but at least it is consisten with
how elements are derefernced in other parts of PTZListModel. The code does need
to be revisited to make sure changes to the number of devices doesn't cause
dereferencing to the wrong device.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>