Reenable SerialPort support
Support for UART connected cameras requires the SerialPort library from
Qt, but the OBS deps package doesn't include that library, which makes
it difficult to enable on Windows and MacOS. However on Linux it is
packaged by the distro. At configure time, check to see if the library
is available. If it is, then enable support for UART connected cameras.

UART support can be explicitly disabled by passing
'-D ENABLE_SERIALPORT=OFF' to cmake at configure time.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>