Fix layout of properties dialog
The OBSPropertiesView widget is supposed to appear above the buttons,
not below them. Fix the layout to always put it in the correct location.
The solution was to add a new horizontal layout object in the right
location rather than relying on a fixed offset into the widget list of
the containing layout.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>