Fix pan/tilt/zoom on slowest speed setting
On VISCA, pan, tilt and zoom were not working when the speed slider was set
to the lowest speed because when calculating the VISCA arguments with
speed=0, the encoding function treated '0' as stop. Adjust the calculation
so that the lowest speed on the slider is non-zero, and so that the slowest
pan, tilt and zoom commands are encoded correctly.

Fixes #101
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>