Windows: Remove old plug-in at install
There have been errors reported after installing OBS Studio v28 and the
latest version of this plug-in that “ptz-controls.dll” failed to load.
This is because the name of the dll has changed, and a copy of the old
name is still hanging around. Solve this bug by removing the old version
at install time.

Fixes #115

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>