Cache movement commands so backend can throttle
Make the `pantilt()`, `zoom()`, and `focus()` methods generic and cache
the chosen speed settings so that the backend driver can easily throttle
commands going out on the wire. Now a new virtual method, `do_update()`
needs to be implemented that notifies the backend that changes have been
made, and that it should send out or schedule a command to the camera.

This change also reworks the VISCA backend to generate
pan/tilt/zoom/focus commands in the `send_pending()` method. If the
VISCA device is idle, then the command gets sent out immediately, but if
the device is already processing a command then it waits until the
current command completes before sending.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>