PTZ Action in preview acting only if ptz cam source is not in program
This patch implements a new PTZ ACTION option called
PTZ_ACTION_TRIGGER_PREVIEW_ONLY_ACTIVE, that blocks ptz actions when
the same cam is found as a cam source in the program scene.

obs-ptz already have the hability to block ptz buttons when the ptz
cam is active in program, now this option creates a similar mecanism
for PTZ Action sources attached to the scenes.

Signed-off-by: Fabio Ferrari <fabio@ferraribr.com>