Keep track of which settings are 'stale'
Add a list of property names that may no longer be accurate. For
example, the `pan_pos` and `tilt_pos` values are not longer accurate
when a pan/tilt movement is made. This patch uses a QSet to keep track
of the stale properties which can later be used to know what inquiry
commands to send.

It also adds an 'affects' field to the PTZCmd object so that commands
can be annotated with the property that it affect

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>