Gamepad Fixes
== Description ==
Put in some null checks to make sure the gamepad doesn't try to access some of the UI when it's not ready. Better cleanup of the thread; let the qt parent handle the memory. Also let the config load before setting up the gamepad, this was the intent so the setting was read first.

Signed-off-by: Eric Schmidt <dragonrhythm@gmail.com>