settings: Move 'apply' button to the interface section
The apply button in the settings dialog should only apply to the
interface section. All of the other settings are intended to be modified
'live'. e.g., changing a setting will have immediate affect. This patch
removes the 'Apply' and 'Close' buttons from the main dialog display and
replaces them with an 'Apply' button in the interface group.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>