visca: Track active commands in 'active_cmd' array
The visca code was implicitly treating the first command in the
`pending_cmds` list to be the active command on the wire, but this isn't
entirely accurate. There may not be any active command, or there may be
a command sent that isn't being tracked in pending_cmds. To better keep
track of what is happening, repurpose the `active_cmd[]` array to store
a copy of the sent command and remove it from the `pending_cmds[]` array
when it gets sent on the wire.

This commit doesn't change the behaviour of command packets getting sent
immediately and unconditionally, whereas inquriries get queued. Cleaning
up that behaviour will be done in a separate patch.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>