VISCA: add new helper for +/-ve zoom & focus
Adds new command encodings for signed zoom and focus moves. This allows
the code to be simplified when generating a command so that it doesn't
need to choose between three commands depending on the sign (-ve move,
+ve move, and stop). Instead a signed value can just be passed in and
the encoding will do the right thing. Just like it does for
pantilt_drive.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>