visca: Keep track of camera power status
Use the STATUS_CONNECTED flag to keep track of whether or not the camera
is responding. If the camera is not responding then clear the flag and
don't attempt to send any inquiry commands. Movement commands will still
be sent to the device so that receive-only cameras can still be used. If
the camera starts responding, then STATUS_CONNECTED will be set back to
true.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>