Fix crash updating OBSPropertiesView in callback
There is a loop where a change to the OBSPropertiesView will call
set_settings(), which emits settingsChanged(), which calls
RefreshProperties() on the OBSPropertiesView. This call loop changes the
data structures and deallocates memory while the callback invocation is
still using them which results in a segfault.

Temporary fix is to remove the recursion. Long term fix will require the
settings dialog to protect itself against this loop while running
within a callback, and a cleaner definition of how chagned data flows
from camera, to PTZDevice object, to the settings dialog.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>