Strictly associate devices with an OBS source
Instead of leaving the PTZ device name as freeform, strictly enforce
that it needs to be associated with a source in order to get autoselect
to work. With this change it is no longer possible to enter any name for
a PTZ device. Instead the user must select from a list of unassigned
sources. If a source is not selected, then a default name is assigned.

This also changes the settings dialog to display the device connection
in the list box instead of just the name. This makes it easier to
identify which device is which based on the connection.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>