settings: Filter unnecessary device renames
If a device isn't associated with a source, then it gets a default name
for "PTZ Device[ #]". However, when there are multiple devices without a
source assigned then selecting one of those devices causes it to be
unnecessarily renamed. Fix the code so that unassigned devices can keep
the generic name it already has.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>