visca: Simplify timeout timer handling
Simplify the timeout timer management by adding a wrapper around
send_immediate() so that the timer can be reset in a consistent way
instead of duplicating code that can get out of sync. Also remove the
nested if statements from timeout() as there are really only two
conditions to test for. Either the current message needs to be retried,
or it's timed out and the next message needs to be generated. This can
be done with a simple if/else block.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>