visca: Fix signed value encoding
Commit 910880, "Fix pan/tilt/zoom on slowest speed setting" attempted to
correctly encode the lowest possible speed for a movement. However, the
calculation was wrong so that when sending a 'stop' command, the speed
was encoded at a very high value. Fix this by explicitly testing if the
value is '0', and encoding the right thing instead of trying to do
something clever.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>