pelco: Fix uncontrolled zoom
Commit 8158a65b, "Cache movement commands so backend can throttle" broke
PELCO zoom control by not sending the correct stop command. The code was
looking for a never satisfied condition (abs(speed) < 0). It also had
the bug that if zoom and pan/tilt were operating at the same time, but
zoom changes would cancel the pan/tilt movement. Fix this by rewriting
the update function to collect the pan, tilt and zoom movements into a
single message.

Fixes: #131
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>