visca: Add configurable range limits
Instead of hard coding camera parameters, provide tuneable properties
that can be changed depending on the camera model. Currently the plugin
doesn't have any way to determine what the parameters should be. The
user needs to look them up in the camera manual. A future improvement
will be to add a calibration mode which can put the camera through a set
of movements and read the resulting position.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>