visca: fix timeouts on unidirectional cameras
VISCA cameras can be wired to receive only on the serial link, which
means the camera never responds to commands or inquiries. However,
commit 9158a65b, "Cache movement commands so backend can throttle" broke
responsive control of these cameras by setting a timeout period of 3
times per second (was intended to be 30 times per second). Fix the
timeout to use the correct value. At the same time, don't try to retry
commands if the camera is not responding. Instead check the connected
flag, and drop the timed out commands quietly until when and if the
camera starts responding.

This behaviour should also work with UDP connected cameras that cannot
send replies back (e.g., if they are on a separate subnet and the
gateway isn't configured)..

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>