visca: fix encoding of pan/tilt speed
The VISCA PanTilt command is encodes speed and direction separately
where direction is a 2 bit field, and speed is a value in the range
0x01-0x??. However the code was encoding the lowest speed as 0 instead
of 1 as documented in the Sony manual. (note: it is possible that other
cameras will accept '0' as the slowest speed). Modify the code to encode
correctly, and use std::clamp() to ensure the value isn't outside the
valid range. As a bonus, this change simplifies the codes overall.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>