Make number of presets configurable
Many cameras have more than 16 presets. Make it possible to have more
than 16 in the preset list. For now this is just a simple tunable that
changes the number of presets that are preloaded into the list. The
preset system does need more rework to make it possible to reorder
presets and have preset data stored locally, but that is a job for a
future commit.

Fixes #86

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>