ui: add touch control widget
This is an experimental interface for providing direct control over
camera movement speed. It creates a new widget than can be switched on
from the bottom toolbar. When visible, it provides a circle with
direction arrows in it for pan/tilt control. Clicking near the center of
the circle provides slow movment. Clicking near the outside edge
provides fast control. Draging the mouse inside the circle changes the
speed during the move.

Renames the buttonResizeFilter to squareResizeFilter so that it can be
used with the new widget also.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>