ui: Teach the list views to wrap around
It is useful for the camera device and preset lists to wrap around when
the top or bottom of the list is reached, but the QListView isn't able
to do that by default. Extend QListView to add the wrap around
behaviour, and update the hotkey code to use the new behaviour for
changing selected device.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>