Fix build issues on Fedora 37
A number of problems were exposed when building on Fedora 37 using
RPMFusion packaged obs-studio. Some implicitly included headers needed
to be added. The touch control needed to use the older ->pos() method
for compatibility, and the byte |= operator needed to be modified.

Also update the documentation to cover building against Qt5.

Fixes: #138, Cannot build the plugin in Fedora 37
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>