Windows: Add libsdl to dependencies
In the ci script, clone a copy of libsdl and build it so that the
Windows releases can have joystick support enabled by default. libsdl is
cloned in the same way as the obs-studio repo is cloned and installed in
the build-dependencies folder so that cmake find_packge() can locate it
in the normal way.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>