Fix drift on joystick
Some joysticks don't center right to 0.0, which causes the camera pan,
tilt and zoom to drift uncontrollably. Add a slider to set a joystick
'deadzone' which blocks out unintentional movements.

This patch also clears any active movements when the joystick control
feature is turned off so that if there is still uncontrollable drift
with the joystick then the user can still regain control.

Fixes #140

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>