onvif: refactor helper classes as PTZOnvif methods
The Onvif*Service helper classes aren't getting in the way as an
abstraction rather than helping. They don't have access to the data that
they need, and so things like user name and device address have to be
passed in each time and the objects get created/destroyed multiple
times. It also makes it difficult to change to asynchronos SOAP calls
because it expects to receive the reply all within the lifetime of the
object, blocking the main UI thread in the process.

Instead of abstracting the different service calls as separate objects,
merge them into PTZOnvif as methods that can use the PTZOnvif data
directly. This is a first step toward making ONVIF command processing
asynchronos.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>