onvif: Throttle pan/tilt/zoom commands
The ONVIF support was blindly sending requests to the camera without
waiting for a response. If the camera is slow, then these requests queue
up and take a long time to clear. The camera will keep making movements
long after the user has stopped sending movement commands.

The simple solution is to set a flag whenever a pan/tilt/zoom command is
sent, and then clear that flag when a response is received. It isn't
perfect, but it works for now. It might be more functional to have
a limited number of pan/tilt/zoom requests in flight, but that can be
implemented in a future commit.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>